+++ tests/helpers.py
@@ -271,7 +271,7 @@ job_args = (
         "job_type": JobType.CYCLIC,
         "timing": [dt.timedelta(hours=1)],
         "handle": foo,
-        "params": None,
+        "kwargs": None,
         "max_attempts": 1,
         "weight": 1,
         "delay": True,
@@ -284,7 +284,7 @@ job_args = (
         "job_type": JobType.MINUTELY,
         "timing": [dt.time(second=20)],
         "handle": bar,
-        "params": {"msg": "foobar"},
+        "kwargs": {"msg": "foobar"},
         "max_attempts": 20,
         "weight": 0,
         "delay": False,
@@ -297,7 +297,7 @@ job_args = (
         "job_type": JobType.DAILY,
         "timing": [dt.time(hour=7, minute=5)],
         "handle": foo,
-        "params": None,
+        "kwargs": None,
         "max_attempts": 7,
         "weight": 1,
         "delay": True,
@@ -311,18 +311,18 @@ job_args = (
 job_reprs = (
     [
         "scheduler.Job(<JobType.CYCLIC: 1>, [datetime.timedelta(seconds=3600)], <function foo at 0x",
-        ">, {}, 1, 1, True, datetime.datetime(2021, 5, 26, 3, 55), None, True, None)",
+        ">, (), {}, 1, 1, True, datetime.datetime(2021, 5, 26, 3, 55), None, True, None)",
     ],
     [
         "scheduler.Job(<JobType.MINUTELY: 2>, [datetime.time(0, 0, 20)], <function bar at 0x",
         (
-            ">, {'msg': 'foobar'}, 20, 0, False, datetime.datetime(2021, 5, 26, 3, 54, 15),"
+            ">, (), {'msg': 'foobar'}, 20, 0, False, datetime.datetime(2021, 5, 26, 3, 54, 15),"
             " datetime.datetime(2021, 5, 26, 4, 5), False, None)"
         ),
     ],
     [
         "scheduler.Job(<JobType.DAILY: 4>, [datetime.time(7, 5)], <function foo at 0x",
-        ">, {}, 7, 1, True, datetime.datetime(2021, 5, 26, 3, 55), None, True, None)",
+        ">, (), {}, 7, 1, True, datetime.datetime(2021, 5, 26, 3, 55), None, True, None)",
     ],
 )
 
@@ -331,7 +331,7 @@ job_args_utc = (
         "job_type": JobType.CYCLIC,
         "timing": [dt.timedelta(hours=1)],
         "handle": foo,
-        "params": None,
+        "kwargs": None,
         "max_attempts": 0,
         "weight": 1 / 3,
         "delay": False,
@@ -344,7 +344,7 @@ job_args_utc = (
         "job_type": JobType.HOURLY,
         "timing": [dt.time(hour=7, minute=5, tzinfo=utc)],
         "handle": print,
-        "params": None,
+        "kwargs": None,
         "max_attempts": 0,
         "weight": 20,
         "delay": False,
@@ -357,7 +357,7 @@ job_args_utc = (
         "job_type": JobType.WEEKLY,
         "timing": [Weekday.MONDAY],
         "handle": bar,
-        "params": None,
+        "kwargs": None,
         "max_attempts": 0,
         "weight": 1,
         "delay": False,
@@ -373,7 +373,7 @@ job_args_utc = (
             (Weekday.TUESDAY, dt.time(23, 45, 59, tzinfo=utc)),
         ],
         "handle": print,
-        "params": {"end": "FOO\n"},
+        "kwargs": {"end": "FOO\n"},
         "max_attempts": 1,
         "weight": 1,
         "delay": True,
@@ -388,14 +388,14 @@ job_reprs_utc = (
     [
         "scheduler.Job(<JobType.CYCLIC: 1>, [datetime.timedelta(seconds=3600)], <function foo at 0x",
         (
-            ">, {}, 0, 0.3333333333333333, False, datetime.datetime(2021, 5, 26, 3, 54, 59, 999990"
+            ">, (), {}, 0, 0.3333333333333333, False, datetime.datetime(2021, 5, 26, 3, 54, 59, 999990"
             ", tzinfo=datetime.timezone.utc), None, True, datetime.timezone.utc)"
         ),
     ],
     [
         (
             "scheduler.Job(<JobType.HOURLY: 3>, [datetime.time(0, 5, tzinfo=datetime.timezone.utc)],"
-            " <built-in function print>, {}, 0, 20, False, datetime.datetime(2021, 5, 26, 3, 55,"
+            " <built-in function print>, (), {}, 0, 20, False, datetime.datetime(2021, 5, 26, 3, 55,"
             " tzinfo=datetime.timezone.utc), datetime.datetime(2021, 5, 26, 23, 55, "
             "tzinfo=datetime.timezone.utc), False, datetime.timezone.utc)"
         )
@@ -403,7 +403,7 @@ job_reprs_utc = (
     [
         "scheduler.Job(<JobType.WEEKLY: 5>, [<Weekday.MONDAY: 0>], <function bar at 0x",
         (
-            ">, {}, 0, 1, False, datetime.datetime(2021, 5, 25, 3, 55, "
+            ">, (), {}, 0, 1, False, datetime.datetime(2021, 5, 25, 3, 55, "
             "tzinfo=datetime.timezone.utc), None, True, datetime.timezone.utc)"
         ),
     ],
@@ -411,7 +411,7 @@ job_reprs_utc = (
         (
             "scheduler.Job(<JobType.WEEKLY: 5>, [<Weekday.WEDNESDAY: 2>, (<Weekday.TUESDAY: 1>,"
             " datetime.time(23, 45, 59, tzinfo=datetime.timezone.utc))], <built-in function print>"
-            ", {'end': 'FOO\\n'}, 1, 1, True, "
+            ", (), {'end': 'FOO\\n'}, 1, 1, True, "
             "datetime.datetime(2021, 6, 2, 3, 55, tzinfo=datetime.timezone.utc),"
             " datetime.datetime(2021, 7, 25, 3, 55, tzinfo=datetime.timezone.utc),"
             " False, datetime.timezone.utc)"